# packages/core/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Computer Interaction API",
    "version": "1.0.0",
    "description": "API to interact with and control a computer.",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4242/{computer_name}",
      "description": "Local development server",
      "variables": {
        "computer_name": {
          "default": "your-computer-name",
          "description": "Computer name"
        }
      }
    }
  ],
...
  "paths": {
    "/computer/screenshot": {
      "get": {
        "tags": ["Computer"],
        "summary": "Take a Screenshot",
        "description": "Captures and returns a screenshot from the specified display as a PNG file.",
        "operationId": "computerTakeScreenshot",
        "parameters": [
          {
            "$ref": "#/components/parameters/DisplayNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Screenshot captured successfully.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "x-implementation-details": "The implementation captures the entire screen of the specified display and returns it as a PNG file."
      }
    },
    "/computer/key": {
      "post": {
        "tags": ["Computer"],
        "summary": "Press key",
        "description": "Presses a key on the specified display.",
        "operationId": "computerPressKey",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "description": "Key to press."
                  },
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to press the key."
                  }
                },
                "required": ["key"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Key pressed successfully."
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/computer/type": {
      "post": {
        "tags": ["Computer"],
        "summary": "Type text",
        "description": "Types text on the specified display.",
        "operationId": "computerTypeText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Text to type."
                  },
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to type text."
                  },
                  "typing_delay": {
                    "type": "integer",
                    "default": 12,
                    "description": "Delay (in milliseconds) between typed characters."
                  },
                  "typing_group_size": {
                    "type": "integer",
                    "default": 50,
                    "description": "Number of characters typed between delays."
                  }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Text typed successfully."
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/computer/mouse-move": {
      "post": {
        "tags": ["Computer"],
        "summary": "Move mouse cursor",
        "description": "Moves the mouse cursor to the specified coordinates on the given display.",
        "operationId": "computerMoveCursor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "x": {
                    "type": "integer",
                    "description": "X coordinate."
                  },
                  "y": {
                    "type": "integer",
                    "description": "Y coordinate."
                  },
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to move the mouse."
                  }
                },
                "required": ["x", "y"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Mouse moved successfully."
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/computer/left-click": {
      "post": {
        "tags": ["Computer"],
        "summary": "Perform left mouse click",
        "description": "Performs a left mouse click on the specified display.",
        "operationId": "computerLeftClick",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to perform the click."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Left click performed successfully."
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/computer/right-click": {
      "post": {
        "tags": ["Computer"],
        "summary": "Perform right mouse click",
        "description": "Performs a right mouse click on the specified display.",
        "operationId": "computerRightClick",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to perform the click."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Right click performed successfully."
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/computer/middle-click": {
      "post": {
        "tags": ["Computer"],
        "summary": "Perform middle mouse click",
        "description": "Performs a middle mouse click on the specified display.",
        "operationId": "computerMiddleClick",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to perform the click."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Middle click performed successfully."
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
    "/computer/double-click": {
      "post": {
        "tags": ["Computer"],
        "summary": "Perform double click",
        "description": "Performs a double click with the left mouse button on the specified display.",
        "operationId": "computerDoubleClick",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to perform the double click."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Double click performed successfully."
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/computer/left-click-drag": {
      "post": {
        "tags": ["Computer"],
        "summary": "Perform left click and drag",
        "description": "Performs a left click and drag operation to the specified coordinates on the given display.",
        "operationId": "computerLeftClickDrag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "x": {
                    "type": "integer",
                    "description": "Target X coordinate to drag to."
                  },
                  "y": {
                    "type": "integer",
                    "description": "Target Y coordinate to drag to."
                  },
                  "display_num": {
                    "type": "integer",
                    "default": 1,
                    "description": "Display number on which to perform the drag operation."
                  }
                },
                "required": ["x", "y"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Drag operation performed successfully."
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/computer/cursor-position": {
      "get": {
        "tags": ["Computer"],
        "summary": "Get cursor position",
        "description": "Gets the current position of the mouse cursor on the specified display.",
        "operationId": "computerGetCursorPosition",
        "parameters": [
          {
            "name": "display_num",
            "in": "query",
            "description": "Display number to get the cursor position from.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor position coordinates.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "x": {
                      "type": "integer"
                    },
                    "y": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/HTTPError"
            }
          }
        }
      },
...
    },
    "parameters": {
      "DisplayNumber": {
        "name": "display_num",
        "in": "query",
        "description": "Display number to operate on, defaults to 1.",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 1
        }
      }
    },
...
  }
}